clean reorganization of the common part
