+++ java/example/common/src/main/java/org/eclipse/keyple/example/common/calypso/postructure/HoplinkInfo.java
@@ -9,7 +9,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
-package org.eclipse.keyple.example.calypso.common.postructure;
+package org.eclipse.keyple.example.common.calypso.postructure;
