+++ java/example/common/src/main/java/org/eclipse/keyple/example/common/generic/SeProtocolDetectionEngine.java
@@ -9,7 +9,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
-package org.eclipse.keyple.example.generic.common;
+package org.eclipse.keyple.example.common.generic;
