+++ java/example/common/src/main/java/org/eclipse/keyple/example/common/generic/stub/StubMifareClassic.java
@@ -9,7 +9,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
-package org.eclipse.keyple.example.generic.pc.stub.se;
+package org.eclipse.keyple.example.common.generic.stub;
 
 
 import org.eclipse.keyple.core.util.ByteArrayUtil;
@@ -18,14 +18,14 @@ import org.eclipse.keyple.plugin.stub.StubSecureElement;
 /**
  * Simple contact stub SE (no command)
  */
-public class StubMemoryST25 extends StubSecureElement {
+public class StubMifareClassic extends StubSecureElement {
 
-    final static String seProtocol = "PROTOCOL_MEMORY_ST25";
-    final String ATR_HEX = "3B8F8001804F0CA000000306070007D0020C00B6";
+    final static String seProtocol = "PROTOCOL_MIFARE_CLASSIC";
+    final String ATR_HEX = "3B8F8001804F0CA000000306030001000000006A";
 
-    public StubMemoryST25() {
+    public StubMifareClassic() {
         /* Get data */
-        addHexCommand("FFCA 000000", "8899AABBCCDDEEFF9000");
+        addHexCommand("FFCA 000000", "112233449000");
     }
 
     @Override
