+++ java/example/common/src/main/java/org/eclipse/keyple/example/common/generic/stub/StubSe2.java
@@ -9,21 +9,21 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
-package org.eclipse.keyple.example.generic.pc.stub.se;
+package org.eclipse.keyple.example.common.generic.stub;
 
 
 import org.eclipse.keyple.core.util.ByteArrayUtil;
 import org.eclipse.keyple.plugin.stub.StubSecureElement;
 
 /**
- * Simple contact stub SE (no command)
+ * Simple contactless stub SE (no command)
  */
-public class StubSe1 extends StubSecureElement {
+public class StubSe2 extends StubSecureElement {
 
-    final static String seProtocol = "PROTOCOL_ISO7816_3";
-    final String ATR_HEX = "3B3F9600805A0080C120000012345678829000";// serial number : 12345678
+    final static String seProtocol = "PROTOCOL_ISO14443_4";
+    final String ATR_HEX = "3B8E800180318066409089120802830190000B";
 
-    public StubSe1() {}
+    public StubSe2() {}
 
     @Override
     public byte[] getATR() {
@@ -34,6 +34,4 @@ public class StubSe1 extends StubSecureElement {
     public String getSeProcotol() {
         return seProtocol;
     }
-
-
 }
